Extract option definitions to own file
